requirements.txt
django
djangorestframework
django-crispy-forms
requests
django-tables2
jsonpickle
gunicorn